refactor: move gravestone-specific items to small gravestone module
