+++ src/main.rs
@@ -111,7 +111,6 @@ fn main() {
             ConditionSet::new()
                 .run_in_state(GameState::LevelTransition)
                 .with_system(gameplay::transitions::spawn_level_card)
-                .with_system(gameplay::transitions::spawn_gravestone_body)
                 .with_system(gameplay::transitions::spawn_control_display)
                 .with_system(gameplay::transitions::load_next_level)
                 .with_system(gameplay::transitions::level_card_update)
@@ -142,10 +141,6 @@ fn main() {
         .add_system(sugar::goal_ghost_animation.run_not_in_state(GameState::AssetLoading))
         .add_system(sugar::goal_ghost_event_sugar.run_not_in_state(GameState::AssetLoading))
         .add_system(sugar::animate_grass_system.run_not_in_state(GameState::AssetLoading))
-        .register_ldtk_entity::<bundles::InputBlockBundle>("W")
-        .register_ldtk_entity::<bundles::InputBlockBundle>("A")
-        .register_ldtk_entity::<bundles::InputBlockBundle>("S")
-        .register_ldtk_entity::<bundles::InputBlockBundle>("D")
         .register_ldtk_entity::<bundles::GoalBundle>("Goal")
         .register_ldtk_entity::<bundles::GrassBundle>("Grass")
         .register_ldtk_int_cell::<bundles::ExorcismBlockBundle>(2)
